# promql/engine: Introduce SelectManager to control execution of .Select requests