Signed-off-by: Kemal Akkoyun <kakkoyun@gmail.com>

This PR attempts to fix https://github.com/prometheus/prometheus/issues/6878

Currently, Prometheus executes PromQL “.Select” queries sequentially. Particularly for the remote storages, this implementation wastes CPU cycles while waiting for independent queries.

Requirements that gathered from the comments of https://github.com/prometheus/prometheus/issues/6878 and https://github.com/prometheus/prometheus/pull/6887.

#### Requirements
* Should NOT use all the available CPU.
  * Limit concurrency
* PromQL should not care about how the querier is implemented, it should always be the same logic.
* Should avoid introducing changes to LOCAL storage.
* Should avoid spawning goroutines if there is just a single selector.

This PR fulfils the requirements listed above by introducing a small component to regulate select requests against querier.

It still misses tests and also I'm gonna add benchmarks.

P.S: Regarding https://github.com/prometheus/prometheus/pull/7089#issuecomment-617132242 after checking out the comments, I decided a design doc might not be needed.

<!--
    Don't forget!
    
    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.
    
    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to actual library usage.
    
    - No tests are needed for internal implementation changes.
    
    - Performance improvements would need a benchmark test to prove it.
    
    - All exposed objects should have a comment.
    
    - All comments should start with a capital letter and end with a full stop.
 -->

### Benchmarks

#### Old
<details>Master
<p>

```
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
BenchmarkRangeQuery/expr=a_one,steps=1-12         	   75193	     17112 ns/op	    5235 B/op	     101 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-12        	   74503	     17711 ns/op	    5235 B/op	     101 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-12       	   50686	     23400 ns/op	    5394 B/op	     103 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-12      	   12404	     80742 ns/op	    7238 B/op	     114 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=1-12         	   13332	     89027 ns/op	   12116 B/op	     195 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=10-12        	   13372	     90231 ns/op	   12116 B/op	     195 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=100-12       	    7586	    166988 ns/op	   12998 B/op	     206 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=1000-12      	    1590	    741547 ns/op	   21344 B/op	     289 allocs/op
BenchmarkRangeQuery/expr=a_hundred,steps=1-12     	    1431	    854308 ns/op	   79258 B/op	    1098 allocs/op
BenchmarkRangeQuery/expr=a_hundred,steps=10-12    	    1293	    877883 ns/op	   79518 B/op	    1098 allocs/op
BenchmarkRangeQuery/expr=a_hundred,steps=100-12   	     759	   1609190 ns/op	   87226 B/op	    1199 allocs/op
BenchmarkRangeQuery/expr=a_hundred,steps=1000-12  	     162	   7309513 ns/op	  160699 B/op	    2002 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-12         	   59626	     19912 ns/op	    6234 B/op	     132 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-12        	   58015	     20682 ns/op	    6236 B/op	     132 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-12       	   33399	     35487 ns/op	    6395 B/op	     134 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-12      	    7819	    162192 ns/op	    8204 B/op	     144 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-12         	   10000	    102998 ns/op	   16344 B/op	     263 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-12        	   10000	    109723 ns/op	   16344 B/op	     263 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-12       	    4804	    257045 ns/op	   17227 B/op	     274 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-12      	     793	   1497840 ns/op	   25053 B/op	     347 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-12     	    1270	    973073 ns/op	  115238 B/op	    1532 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-12    	    1137	   1031542 ns/op	  115289 B/op	    1532 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-12   	     474	   2519348 ns/op	  123282 B/op	    1633 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-12  	      81	  14960258 ns/op	  192899 B/op	    2337 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-12     	     771	   1515496 ns/op	   40897 B/op	     273 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-12     	      82	  15242966 ns/op	  155918 B/op	    1306 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-12 	       7	 153293705 ns/op	 1301545 B/op	   11669 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12         	    1375	    867148 ns/op	 1178077 B/op	     270 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12        	     897	   1369617 ns/op	 1179225 B/op	     270 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12       	     199	   6211680 ns/op	 1180731 B/op	     271 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12      	      22	  52517652 ns/op	 1195297 B/op	     278 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12         	     180	   6701682 ns/op	 1247882 B/op	    1061 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12        	     100	  11082399 ns/op	 1249110 B/op	    1058 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12       	      20	  57866242 ns/op	 1227512 B/op	    1065 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12      	       2	 542526223 ns/op	 1279420 B/op	    1141 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12     	      18	  65587749 ns/op	 1933669 B/op	    8901 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12    	      10	 112111123 ns/op	 1935263 B/op	    8898 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12   	       2	 583285199 ns/op	 1952756 B/op	    9005 allocs/op
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12  	       1	5093915142 ns/op	 4953760 B/op	   10061 allocs/op
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-12                        	    1728	    685480 ns/op	  548533 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-12                       	    1450	    837450 ns/op	  548793 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-12                      	     500	   2397921 ns/op	  549249 B/op	     223 allocs/op
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-12                     	      64	  17995964 ns/op	  557419 B/op	     230 allocs/op
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-12                        	     199	   5945700 ns/op	  620237 B/op	    1010 allocs/op
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-12                       	     160	   7683115 ns/op	  620504 B/op	    1010 allocs/op
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-12                      	      51	  22780568 ns/op	  631275 B/op	    1021 allocs/op
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-12                     	       6	 179553652 ns/op	  740890 B/op	    1098 allocs/op
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-12                    	      19	  58537583 ns/op	 1338492 B/op	    8851 allocs/op
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-12                   	      15	  73250833 ns/op	 1342248 B/op	    8851 allocs/op
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-12                  	       5	 228778356 ns/op	 1429876 B/op	    8955 allocs/op
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-12                 	       1	1712454614 ns/op	 4703576 B/op	    9956 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-12                           	    1608	    660447 ns/op	  548736 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-12                          	    1594	    756493 ns/op	  548283 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-12                         	     698	   1723085 ns/op	  549081 B/op	     223 allocs/op
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-12                        	     100	  11224424 ns/op	  555193 B/op	     230 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-12                           	     205	   5753162 ns/op	  616240 B/op	    1010 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-12                          	     180	   6672029 ns/op	  620452 B/op	    1010 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-12                         	      78	  16072713 ns/op	  632859 B/op	    1020 allocs/op
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-12                        	      10	 110484667 ns/op	  624317 B/op	    1091 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-12                       	      21	  57354263 ns/op	 1302662 B/op	    8849 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-12                      	      18	  66431425 ns/op	 1305330 B/op	    8851 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-12                     	       7	 160115955 ns/op	 1315096 B/op	    8957 allocs/op
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-12                    	       1	1106999909 ns/op	 4718664 B/op	    9959 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-12               	    1812	    644683 ns/op	  548963 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-12              	    1803	    672138 ns/op	  548655 B/op	     222 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-12             	    1411	    854606 ns/op	  550780 B/op	     223 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-12            	     441	   2567926 ns/op	  565235 B/op	     230 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-12               	     214	   5675795 ns/op	  616283 B/op	    1009 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-12              	     208	   5742605 ns/op	  617538 B/op	    1009 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-12             	     158	   7548893 ns/op	  642463 B/op	    1019 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-12            	      48	  24766373 ns/op	  801943 B/op	    1089 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-12           	      21	  54814914 ns/op	 1302817 B/op	    8842 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-12          	      20	  56648193 ns/op	 1348786 B/op	    8843 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-12         	      15	  73846118 ns/op	 1532006 B/op	    8945 allocs/op
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-12        	       5	 265731611 ns/op	 3147916 B/op	    9648 allocs/op
BenchmarkRangeQuery/expr=-a_one,steps=1-12                                    	   69477	     16949 ns/op	    5858 B/op	     113 allocs/op
BenchmarkRangeQuery/expr=-a_one,steps=10-12                                   	   69741	     16884 ns/op	    5858 B/op	     113 allocs/op
BenchmarkRangeQuery/expr=-a_one,steps=100-12                                  	   48648	     24271 ns/op	    6019 B/op	     115 allocs/op
BenchmarkRangeQuery/expr=-a_one,steps=1000-12                                 	   14346	     83263 ns/op	    7861 B/op	     126 allocs/op
BenchmarkRangeQuery/expr=-a_ten,steps=1-12                                    	   12710	     92901 ns/op	   15958 B/op	     244 allocs/op
BenchmarkRangeQuery/expr=-a_ten,steps=10-12                                   	   12649	     94328 ns/op	   15959 B/op	     244 allocs/op
BenchmarkRangeQuery/expr=-a_ten,steps=100-12                                  	    5576	    195339 ns/op	   16842 B/op	     255 allocs/op
BenchmarkRangeQuery/expr=-a_ten,steps=1000-12                                 	    1195	    953016 ns/op	   25159 B/op	     338 allocs/op
BenchmarkRangeQuery/expr=-a_hundred,steps=1-12                                	    1322	    906496 ns/op	  114733 B/op	    1513 allocs/op
BenchmarkRangeQuery/expr=-a_hundred,steps=10-12                               	    1284	    934493 ns/op	  114722 B/op	    1513 allocs/op
BenchmarkRangeQuery/expr=-a_hundred,steps=100-12                              	     711	   1654612 ns/op	  122818 B/op	    1614 allocs/op
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-12                             	     162	   7413050 ns/op	  196071 B/op	    2417 allocs/op
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-12                             	   39568	     29872 ns/op	   10603 B/op	     177 allocs/op
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-12                            	   34824	     35231 ns/op	   11755 B/op	     213 allocs/op
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-12                           	   12328	     98679 ns/op	   23598 B/op	     577 allocs/op
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-12                          	    1797	    690077 ns/op	  142490 B/op	    4199 allocs/op
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-12                             	    6111	    196129 ns/op	   34075 B/op	     418 allocs/op
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-12                            	    4671	    241691 ns/op	   36810 B/op	     465 allocs/op
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-12                           	    1467	    836998 ns/op	   66053 B/op	     949 allocs/op
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-12                          	     186	   6439008 ns/op	  353251 B/op	    5724 allocs/op
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-12                     	     626	   1894298 ns/op	  273610 B/op	    2731 allocs/op
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-12                    	     506	   2390358 ns/op	  297081 B/op	    2856 allocs/op
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-12                   	     136	   9055245 ns/op	  544205 B/op	    4288 allocs/op
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-12                  	      15	  72868190 ns/op	 3011253 B/op	   18161 allocs/op
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-12                         	     177	   6682100 ns/op	 1360576 B/op	   40449 allocs/op
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-12                         	      18	  63847198 ns/op	 3460563 B/op	   53916 allocs/op
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-12                 	       2	 705876462 ns/op	28140180 B/op	  160997 allocs/op
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-12            	   19167	     61298 ns/op	   18611 B/op	     259 allocs/op
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-12           	   18962	     64943 ns/op	   19761 B/op	     295 allocs/op
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-12          	   12788	     98916 ns/op	   31442 B/op	     657 allocs/op
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-12         	    2928	    368821 ns/op	  148577 B/op	    4268 allocs/op
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-12            	    6114	    186729 ns/op	   33838 B/op	     423 allocs/op
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-12           	    5203	    208786 ns/op	   35017 B/op	     459 allocs/op
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-12          	    2397	    510641 ns/op	   47968 B/op	     836 allocs/op
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-12         	     374	   3324862 ns/op	  176376 B/op	    4562 allocs/op
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-12    	     786	   1430395 ns/op	  176319 B/op	    1873 allocs/op
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-12   	     675	   1660998 ns/op	  197480 B/op	    2051 allocs/op
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-12  	     232	   5034437 ns/op	  422543 B/op	    3986 allocs/op
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-12 	      32	  36894843 ns/op	 2662425 B/op	   23003 allocs/op
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-12             	   19612	     61186 ns/op	   18683 B/op	     261 allocs/op
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-12            	   18512	     65295 ns/op	   19840 B/op	     297 allocs/op
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-12           	   10000	    105099 ns/op	   31543 B/op	     659 allocs/op
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-12          	    2547	    500336 ns/op	  148696 B/op	    4270 allocs/op
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-12             	    6331	    191200 ns/op	   34920 B/op	     432 allocs/op
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-12            	    5515	    225364 ns/op	   38984 B/op	     488 allocs/op
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-12           	    1982	    626781 ns/op	   81320 B/op	    1060 allocs/op
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-12          	     266	   4803318 ns/op	  503720 B/op	    6739 allocs/op
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-12     	     804	   1468828 ns/op	  186059 B/op	    1943 allocs/op
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-12    	     676	   1830560 ns/op	  229726 B/op	    2202 allocs/op
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-12   	     177	   6209363 ns/op	  679296 B/op	    4920 allocs/op
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-12  	      22	  49143078 ns/op	 5171197 B/op	   31890 allocs/op
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-12         	   19299	     62572 ns/op	   18684 B/op	     261 allocs/op
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-12        	   18265	     65251 ns/op	   19842 B/op	     297 allocs/op
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-12       	   10000	    109585 ns/op	   31535 B/op	     659 allocs/op
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-12      	    2090	    493761 ns/op	  148687 B/op	    4270 allocs/op
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-12         	    5968	    187582 ns/op	   33841 B/op	     423 allocs/op
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-12        	    5698	    205004 ns/op	   34993 B/op	     459 allocs/op
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-12       	    2446	    494279 ns/op	   47882 B/op	     836 allocs/op
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-12      	     358	   3271752 ns/op	  176232 B/op	    4562 allocs/op
…