@brian-brazil 2020-05-05
> Only after this we can start discussions if we should specify higher than 1 for local TSDB or not.

Having more than 1 concurrent Select for a given query should not be possible against the local TSDB (even if allowed against remotes), as it makes CPU unpredictable due to causing spikes during the select stage. Right now queries can't use more than 1 core each - if you remove that provisioning CPU is now much harder and in addition query throughput performance will be worse. There's also likely an adverse memory impact.

Keep in mind that Select is the fast part of a query generally, so we'd making life harder for users without gaining them much.