@bwplotka 2020-05-05
Awesome, let's keep it 1 for now then on TSDB, we are definitely seeing slowdowns of select so looking forward to concurrency here (: