@kakkoyun 2020-05-05
@brian-brazil @bwplotka 
With the latest changes, Select concurrency configurable for the engine. By default, it will keep existing behaviour for Prometheus. This will enable downstream projects to configure engine for their needs.

For any reason, if we want to enable concurrency for remote read, we need to introduce additional concurrency limits for tsdb select.

That being said, this is ready for review. I have also attached benchmarks.